+++ libcaf_core/caf/config_option.hpp
@@ -19,6 +19,21 @@ namespace caf {
 class CAF_CORE_EXPORT config_option {
 public:
   // -- member types -----------------------------------------------------------
+  /// An iterator over CLI arguments.
+  using argument_iterator = std::vector<std::string>::const_iterator;
+
+  /// Stores the result of a find operation. The option sets `begin == end`
+  /// if the operation could not find a match.
+  struct find_result {
+    /// The begin of the matched range.
+    argument_iterator begin;
+
+    /// The end of the matched range.
+    argument_iterator end;
+
+    /// The value for the config option.
+    std::string_view value;
+  };
 
   /// Custom vtable-like struct for delegating to type-specific functions and
   /// storing type-specific information shared by several config options.
@@ -92,6 +107,10 @@ public:
   /// Returns whether the category is optional for CLI options.
   bool has_flat_cli_name() const noexcept;
 
+  /// Tries to find this option by its long name in `[first, last)`.
+  find_result find_by_long_name(argument_iterator first,
+                                argument_iterator last) const noexcept;
+
 private:
   std::string_view buf_slice(size_t from, size_t to) const noexcept;
 
@@ -103,14 +122,5 @@ private:
   const meta_state* meta_;
   mutable void* value_;
 };
-
-/// Finds `config_option` string with a matching long name in the argument list
-/// [`first`, `last`). Returns an `iterator` to the matching location and a
-/// `string_view` of the option value if the entry is found, or a `iterator`
-/// to `last` with an empty `string_view` otherwise.
-std::pair<std::vector<std::string>::const_iterator, std::string_view>
-find_by_long_name(const config_option& x,
-                  std::vector<std::string>::const_iterator first,
-                  std::vector<std::string>::const_iterator last);
 
 } // namespace caf