+++ libcaf_core/test/actor_system_config.cpp
@@ -149,21 +149,6 @@ CAF_TEST(file input overrides defaults but CLI args always win) {
   CHECK_EQ(content(cfg), res);
 }
 
-CAF_TEST(parsing - with config file cli option) {
-  MESSAGE("Try opening non-existent config file");
-  cfg.clear();
-  auto err = cfg.parse(string_list{"--config-file=test-me"});
-  CHECK_EQ(err, caf::sec::cannot_open_file);
-  CHECK(cfg.remainder.empty());
-  CHECK_EQ(get_or(cfg, "config-file", ""), "test-me");
-
-  cfg.clear();
-  err = cfg.parse(string_list{"--config-file", "test-me"});
-  CHECK_EQ(err, caf::sec::cannot_open_file);
-  CHECK(cfg.remainder.empty());
-  CHECK_EQ(get_or(cfg, "config-file", ""), "test-me");
-}
-
 // Checks whether both a synced variable and the corresponding entry in
 // content(cfg) are equal to `value`.
 #define CHECK_SYNCED(var, ...)                                                 \